requirements.txt
# Core dependencies
pypdf>=4.0.0
ragas>=0.1.7
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.1.0
tqdm>=4.66.0
jinja2>=3.1.0

# Embedding & Vector DB
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Vietnamese embedding model
# pip install dangvantuan/vietnamese-embedding (from GitHub)
# Or: pip install sentence-transformers + load model directly

# LLM providers (for Agent and Judge)
groq>=0.4.0

# Optional: for additional LLM providers
# anthropic>=0.18.0  # Claude
# google-genai>=0.3.0  # Gemini

# Utilities
numpy>=1.24.0
aiofiles>=23.0.0